refactor(desktop): move the settings section registry out of the sheet (#803)
`settingsContract.ts` and PRD #803's Success Criteria both claim that adding
a section requires touching neither the store nor the sheet. That was false:
`SETTINGS_SECTIONS` lived inside `components/SettingsSheet.tsx`, so every
dependent had to edit a #803-owned rendering component to register, and #741
and #802 would have collided on adjacent lines of the same array.

The registry now lives in `lib/settingsRegistry.ts`, beside the contract it
implements. `SettingsSheet.tsx` imports it and becomes a pure rendering
component a dependent never opens. The `sections` prop seam and both of its
layout tests are unchanged; only the import moved.
