+++ docs/develop/desktop-gui.md
@@ -151,7 +151,7 @@ Everything in the store is deliberately about *this installation on this machine
 Three steps, and the contract is deliberately this small. Adding a **setting** is step 1 plus an edit to your own panel; adding a **section** is one registry row and one component. Neither requires touching the store, the sheet, or anything belonging to another feature.
 
 1. **Storage.** Add a field to your feature's section struct in `desktop/src-tauri/src/settings.rs`, or add a new section struct plus one line on `DesktopSettings`. `#[serde(default)]` throughout gives it a default; there is deliberately no `deny_unknown_fields`, so a field a newer build writes survives an older build reading the file. Field names stay `snake_case` and single-word where that is natural: the same struct is serialised to TOML for the user *and* to JSON for the webview, and every name today is one word, so the two agree byte for byte. The first genuinely multi-word field is the point at which a separate webview DTO has to be introduced — not a `rename_all` on this struct, which would make the TOML read badly.
-2. **UI.** Add one row — `id`, `label`, `icon`, `component` — to `SETTINGS_SECTIONS` in `desktop/src/components/SettingsSheet.tsx`, and own your panel component. Implement `SettingsPanelProps` from `desktop/src/lib/settingsContract.ts`: that file *is* the type to build against, and it is short on purpose. A panel is an ordinary React component and owns its own layout — there is no generic key/value renderer, because #741's endpoint list and #802's model manager are not key/value widgets and a renderer built to fit both would fit neither. Send the **whole document** to `onSave` (spread `settings`, replace your own section) so a save can never drop a section this build's UI has not loaded.
+2. **UI.** Add one row — `id`, `label`, `icon`, `component` — to `SETTINGS_SECTIONS` in `desktop/src/lib/settingsRegistry.ts`, and own your panel component. The registry is a module of its own rather than a constant inside the sheet, so registering never means editing a #803-owned rendering component and two dependents cannot collide on adjacent lines of one array; `SettingsSheet.tsx` reads it and is otherwise none of your business. Implement `SettingsPanelProps` from `desktop/src/lib/settingsContract.ts`: that file *is* the type to build against, and it is short on purpose. A panel is an ordinary React component and owns its own layout — there is no generic key/value renderer, because #741's endpoint list and #802's model manager are not key/value widgets and a renderer built to fit both would fit neither. Send the **whole document** to `onSave` (spread `settings`, replace your own section) so a save can never drop a section this build's UI has not loaded.
 3. **Secrets never go in the document, and never in `localStorage`.** The document may hold a non-secret *reference* — which backend holds the key, or a boolean saying one is stored — and nothing more. A real credential belongs behind the `SecretStore` seam PRD #803 names (store/load/delete keyed by a stable identifier, OS keychain as the intended implementation; #802 picks the backend). This is enforced, not advised: `no_settings_key_may_look_like_a_secret` in `settings.rs` fails the build on a key name containing `key`, `token`, `secret`, `password` or `credential`, with a short allowlist for genuine references. If your name really is a reference, add it to `SECRETISH_ALLOWED` with a comment saying which of the two allowed forms it is.
 
 Two things to expect while doing it. `default_document_shape_is_pinned` asserts the exact serialised default document, so **a new field fails that test by design** — updating it is the deliberate act that puts the ownership question in front of a reviewer. And the settings sheet **drops its section column while there is only one section**, rendering that panel full width; your registry row is what brings the column back, with no layout work on your side. Both directions are pinned with stub sections in `desktop/src/components/SettingsSheet.test.tsx`.
